Tests/LibWeb/Text/expected/IntersectionObserver/scroll-margin-expands-root-scrollport.txt

no-margin: intersecting=false ratio=0
with-margin: intersecting=true ratio=1
